+++ Engine_Master/Application.h
@@ -15,7 +15,9 @@ class ModuleTexture;
 class ModuleEditor;
 class ModuleEditorCamera;
 class ModuleModelLoader;
+class ModuleTime;
 class ModuleScene;
+class ModuleDebugDraw;
 
 class Application
 {
@@ -39,6 +41,8 @@ public:
 	ModuleEditorCamera* editorCamera = nullptr;
 	ModuleModelLoader* modelLoader = nullptr;
 	ModuleScene* scene = nullptr;
+	ModuleTime* time = nullptr;
+	ModuleDebugDraw* debugDraw = nullptr;
 
 private:
 	std::list<Module*> modules;