+++ Engine_Master/ModuleEditor.h
@@ -20,6 +20,7 @@ public:
 	~ModuleEditor();
 
 	bool Init();
+	bool Start();
 	update_status PreUpdate();
 	update_status Update();
 	update_status PostUpdate();
@@ -28,9 +29,17 @@ public:
 	void processEvent(SDL_Event);
 
 public:
-	bool scrollToBottom, show_demo_window, show_log_window, show_about_window, show_configuration_window;
+	bool scrollToBottom, show_about_window, show_configuration_window;
 	ImGuiTextBuffer myBuffer;
 	std::vector<float> fps_log;
+
+private:
+	void loadIcons();
+	void drawMainMenu();
+	void drawHierarchyPanel();
+	void drawCameraPanel();
+	void drawInspectorPanel();
+	void drawLogPanel();
 };
 
 #endif __ModuleEditor_H__