+++ Engine_Master/ModuleEditorCamera.h
@@ -6,14 +6,14 @@
 #include "Application.h"
 #include "ModuleWindow.h"
 #include "ModuleInput.h"
-
+#include "ComponentCamera.h"
 #include "Geometry/Frustum.h"
 #include "Math/MathAll.h"
 #include "SDL.h"
 
 
-#define CAM_SPEED 2.5f
-#define CAM_ROTATION_SPEED 1.36f //80 deg
+#define CAM_SPEED 12.5f
+#define CAM_ROTATION_SPEED 2.7f //160 deg
 
 enum MovementMode
 {
@@ -21,13 +21,15 @@ enum MovementMode
 	NONE,
 	FREE,
 	ORBIT,
-
+	COUNT
 };
 
 class ModuleEditorCamera :
 	public Module
 {
 public:
+	ComponentCamera* cam;
+
 	ModuleEditorCamera();
 	~ModuleEditorCamera();
 
@@ -40,9 +42,6 @@ public:
 
 	bool CleanUp();
 
-	void SetFOV(float);
-	void SetAspectRatio(float);
-	void SetPlaneDistances(float, float);
 	void SetPosition(float, float, float);
 	void updatePosition(float);
 	void updateRotation(float);
@@ -51,20 +50,12 @@ public:
 	void updateNavModes();
 	void LookAt(float3);
 
-	float4x4 viewMatrix;
-	float4x4 projectionMatrix;
-
-	Frustum frustum;
-
 private:
 
 	MovementMode navigationMode = NONE;
 	bool isFastMode = false;
 	float orbitAngleX = 0.0f;
 	float orbitAngleY = 0.0f;
-
-
-	void reloadMatrices();
 
 	float getCamSpeed();
 
@@ -75,8 +66,6 @@ private:
 	void moveForward(float, float speed = 0.0f);
 	void moveBackwards(float, float speed = 0.0f);
 
-	void pitch(float, float);
-	void yaw(float, float);
 	void orbitX(float, float3);
 	void orbitY(float, float3);