+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+* Space partitioning
+* Debug draw correctly implemented
+* Readme updated
+* Camera now renders on a imgui window, has skybox and detects whether a mesh is within or withouu to stop rendering it
+* ModuleTime
 * Fix on rotation deformation
 * Changelog created